==> 00-parse_xml_dumb.ipynb <==
import xml.etree.ElementTree as etree
import codecs
import csv
import time
import os
import pandas as pd

pd.set_option("display.max_colwidth", 150)

# Parsing xml
parsing xml can be a difficult, I used this video as a referenace
https://www.youtube.com/watch?v=AeRN4zI7Dhk&t=725s

wiki_xml = "./data/leagueoflegends_pages_current.xml"
enc = "utf-8"
parsed_path = "./data/parsed_lol_data.csv"

# get an overview of how xml files are
# and the namespace key for different types of pages

with open(wiki_xml) as f:
    for x in range(70):
        print(f.readline().strip())

# Nicely formatted time string
def hms_string(sec_elapsed):
    h = int(sec_elapsed / (60 * 60))
    m = int((sec_elapsed % (60 * 60)) / 60)
    s = sec_elapsed % 60
    return "{}:{:>02}:{:>05.2f}".format(h, m, s)

def strip_tag_name(t):
    t = elem.tag
    idx = k = t.rfind("}")
    if idx != -1:
        t = t[idx + 1:]
    return t

totalCount = 0
title = None
start_time = time.time()

title = ''
id = -99
inrevision = False
ns = 0
text = ''
with codecs.open(parsed_path, "w", enc) as parsed_file:
    articlesWriter = csv.writer(parsed_file, quoting=csv.QUOTE_MINIMAL)

    articlesWriter.writerow(['id', 'title', "ns", "text"])
    
    for event, elem in etree.iterparse(wiki_xml, events=('start', 'end')):
        tname = strip_tag_name(elem.tag)
        if event == 'start':
            if tname == 'page': # reset params every page
                title = ''
                id = -99
                inrevision = False
                ns = 0
                text = ''
            elif tname == 'revision':
                # Do not pick up on revision id's
                inrevision = True
            elif tname == 'title':
                title = elem.text
            elif tname == 'id' and not inrevision and elem.text!=None:
                id = int(elem.text)
            elif tname == 'ns' and elem.text!=None:
                ns = int(elem.text)
            elif tname == 'text' and elem.text!=None:
                text = elem.text.replace("\n", "").strip()
        elif (event == "end") & (tname == 'text') & (text == "") & (elem.text!=None): # some text only appear in end
            text = elem.text.replace("\n", " ").strip()
        elif tname == 'page':
            totalCount += 1
            articlesWriter.writerow([id, title, ns, text])
            if totalCount > 1 and (totalCount % 100000) == 0:
                print("{:,}".format(totalCount))
        elem.clear()
        
time_took = time.time() - start_time
print(f"Total runtime: {hms_string(time_took)}")

# Cleaning

df = pd.read_csv(parsed_path)

df

df[df["title"].str.contains(".*Yasuo.*", na=False)]

From looking into the namespaces keys, we can remove lot's of pages based on their namespace,
those pages don't have lore data, most of them are blogs, forums, wiki metadata etc

df["ns"].value_counts()

df[df["ns"] == 710]

# remove based on name space
df = df[~df["ns"].isin([-2,-1,1,2,3,4,5,6,7,8,9,10,11,12,13,15,110,111,114,115,420,421,500,501,502,503,828,829,1200,1201,1202,2000,2001,2002,2300,2301,2302,2900,2901])]

# remove those rows
display(df[df["text"].str.contains("#REDIRECT.* |#redirect.*", na=False)])
df = df[~df["text"].str.contains("#REDIRECT.* |#redirect.*", na=False)] 

# remove anything with title talk
display(df[df["title"].str.contains("Talk:.*|talk:.*", na=False)])
df = df[~df["title"].str.contains("Talk:.*|talk:.*", na=False)]

# remove first few rows as they are related to wiki not lol
display(df.iloc[:20])
df = df.iloc[20:]

df

df.to_csv("data/parsed_cleaned_data.csv", index=False)
