==> 01-preprocess_data.ipynb <==
import pandas as pd
import gc
import re
import textwrap
pd.set_option("display.max_colwidth", 200)

# Statistics about the data

cleaned_data_path = "data/parsed_cleaned_data.csv"

df = pd.read_csv(cleaned_data_path)
df.sample(10)

df["text"] = df["text"].astype(str)
df["title"] = df["title"].astype(str)

# stats about letters count
df["text"].str.len().describe()

# stats about word count
df["text"].str.split().str.len().describe()

df["text"].str.split().str.len().sort_values(ascending=False).head(10)

# clean text and remove html and xml tags

def clean_text(text):

    # get champ before remving {{}}
    # text = re.sub(r'\{\{[^|]*\|([^}]*)\}\}', r'\1', text)

    # Remove {{}} formatting
    text = re.sub(r"\{\{(.*?)\}\}", r"\1", text)
    
    # remove brackets [[item]]
    # text = re.sub(r'\[+([^[\]]+)\]+', r'\1', text)

    # Remove [[]] links
    # text = re.sub(r'\[\[[^]]*\]\]', '', text)
    text = re.sub(r"<ref>.*?</ref>", "", text)
    
    # remove links
    # text = re.sub(r"http[s]?://\S+\s", "", text)

    # Remove <> tags
    text = re.sub(r'<[^>]*>', '', text)
    
    # Remove File: references
    text = re.sub(r'\[\[File:[^\]]*\]\]', '', text)

    # Remove single brackets []
    text = re.sub(r'\[[^]]*\]', '', text)

    # Remove __TOC__ and similar
    text = re.sub(r'__[A-Z]+__', '', text)

    # Remove asterisks
    text = re.sub(r'\*', '', text)
    
    # Remove multiple spaces and newlines
    text = re.sub(r'\s+', ' ', text)
    
    # Remove equal signs
    pattern = r'\=+([^\=]*?)\=+'
    # Replace the matched pattern
    text = re.sub(pattern, r'\1:', text)
    
    # Remove trailing "]"
    text = re.sub(r'[\] ]+$', '', text)
    
    # Trim leading and trailing whitespace
    text = text.strip()

    
    return text

df["full_text"] = df["text"]
df["text"] = df["text"].apply(clean_text)


def split_text_to_multiple_rows(df, text_column, id_column, word_limit, additional_columns):
    new_rows = []
    
    for _, row in df.iterrows():
        text = row[text_column]
        id_prefix = row[id_column]
        words = text.split()
        
        # Split the text into chunks of 'word_limit' words
        chunks = [words[i:i + word_limit] for i in range(0, len(words), word_limit)]
        
        # Create new rows with updated ids and same values for additional columns
        for idx, chunk in enumerate(chunks):
            new_id = f"{id_prefix}-{idx:03d}"
            new_text = ' '.join(chunk)
            new_row = {id_column: new_id, text_column: new_text}
            for col in additional_columns:
                new_row[col] = row[col]
            new_rows.append(new_row)
    
    # Create a new DataFrame from the list of new rows
    new_df = pd.DataFrame(new_rows)
    
    return new_df


# word_limit depend on which model you are using
# optimal chunk size is 1024 token and tokens are about 0.75 word
df = split_text_to_multiple_rows(df, 'text', 'id', word_limit=int(1024*0.75), additional_columns=['title', 'ns', "full_text"])
df

# Remove rows with text less than 14 words
display(df[df["text"].str.split().str.len() <= 14])
df = df[df["text"].str.split().str.len() >= 14]

df.drop(columns="full_text", inplace=True)
df.to_csv("./data/chunks_data.csv",index=False)
