==> 02-embed_chunks.ipynb <==
import pandas as pd
pd.set_option("display.max_colwidth", 200)

# Embedding mode
there are lots of embedding models, you can see which one is performing good on HF leaderboard and suits your needs
https://huggingface.co/spaces/mteb/leaderboard


the max token is important for our chunking word limit, token are about 0.75 word

# Read data
df = pd.read_csv("./data/chunks_data.csv")
df["text"] = df["text"].astype(str)
df["title"] = df["title"].astype(str)
df

input = "This chunk is about " + df['title']  + f" with ID-" + df['id'].astype(str) + ": "  + df["text"]

input = input.astype(str)
input

from sentence_transformers import SentenceTransformer
from sentence_transformers.util import cos_sim
import torch
import torch.nn.functional as F

device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')

model = SentenceTransformer('Alibaba-NLP/gte-base-en-v1.5', trust_remote_code=True, model_kwargs={"torch_dtype":torch.bfloat16})

embeddings = model.encode(input, batch_size=1, show_progress_bar=True)
embeddings_torch = torch.from_numpy(embeddings).to(device)

embeddings.shape

text = "Kayn weapon name?"

text_embeddings = model.encode(text, batch_size=1, show_progress_bar=True)
text_embeddings_torch = torch.from_numpy(text_embeddings).to(device)

cos_sim_score = cos_sim(a=text_embeddings_torch, b=embeddings_torch)[0]

top_result = torch.topk(cos_sim_score, k=5)
top_result

import textwrap

for i in top_result[1]:
    print("title: " + df.iloc[int(i)]["title"] + ": " )
    print(textwrap.fill(df.iloc[int(i)]["text"], width=70))
    print("___________________________\n\n")

torch.save(embeddings_torch, "./data/embeddings_torch.pt")
