==> 03-pure_LLM.ipynb <==
# Model used
I'll be using llama.cpp so I need a model in GGUF format,
huggingface has lots of models and you can filter with GGUF tag

the model I will be using is 
meta-llama-3-8b-instruct.Q4_K_S.gguf which is a quantized model of llama-3-7b 
from https://huggingface.co/SanctumAI/Meta-Llama-3-8B-Instruct-GGUF 

Uncomment the line below to download it, or use huggingface hub through python api

# mkdir models
# !curl -o ./models https://huggingface.co/SanctumAI/Meta-Llama-3-8B-Instruct-GGUF/blob/main/meta-llama-3-8b-instruct.Q4_K_M.gguf 

# llm = Llama.from_pretrained(
#     repo_id="SanctumAI/Meta-Llama-3-8B-Instruct-GGUF",
#     filename="*Q4_k_M.gguf",
#     verbose=True
# )

## Imports

from llama_cpp import Llama
import textwrap
import pandas as pd
from sentence_transformers import SentenceTransformer
from sentence_transformers.util import cos_sim
import torch
import gc
import transformers
import os
from dotenv import load_dotenv

load_dotenv()


# LLM using llama.cpp without RAG

llm = Llama(
    model_path="./models/meta-llama-3-8b-instruct.Q4_K_M.gguf",
    n_gpu_layers=-1,
    n_ctx=4096,
    flash_attn=True

)

output = llm(
    "Who was Zed main rival in Runeterra?",
      max_tokens=None, # set to None to generate up to the end of the context window
      stop=["Q:"], # Stop generating just before the model would generate a new question
      echo=False,
      temperature=0.75
) # Generate a completion, can also call create_completion
# print(output)
print(textwrap.fill(output["choices"][0]["text"], width=70))

del llm
_ = gc.collect()
